feat: equity different tokens and methods, init and restructuring
