+++ package.json
@@ -46,6 +46,7 @@
 		"react-datepicker": "^6.1.0",
 		"react-dom": "^18.2.0",
 		"react-redux": "^9.1.2",
+		"react-select": "^5.8.0",
 		"react-toastify": "^9.1.3",
 		"typechain": "^8.1.1",
 		"viem": "^2.13.3",
