+++ pages/challenges/index.tsx
@@ -9,10 +9,6 @@ export default function ChallengesPage({}) {
 				<title>Frankencoin - Auctions</title>
 			</Head>
 
-{/* 			<div>
-				<AppPageHeader title="All Active Challenges" />
-			</div> */}
-
 			<div className="md:mt-8">
 				<ChallengesTable />
 			</div>
